revert: "feat: widget tool box on mouse hover and selection state"
This reverts commit 1937d7cd774da396067ed90de32c71cfac3754c8.
