+++ packages/dashboard/src/components/widgets/selectionBoxAnchor.tsx
@@ -5,8 +5,6 @@ import {
 } from '../internalDashboard/gestures/determineTargetGestures';
 import './selectionBoxAnchor.css';
 import type { Anchor } from '~/store/actions';
-import WidgetActions from './widgetActions';
-import { useSelectedWidgets } from '~/hooks/useSelectedWidgets';
 
 const CORNERS: Anchor[] = [
   'top-left',
@@ -21,9 +19,6 @@ export type SelectionBoxAnchorProps = {
 };
 
 const SelectionBoxAnchor: React.FC<SelectionBoxAnchorProps> = ({ anchor }) => {
-  const selectedWidgets = useSelectedWidgets();
-  const widget = selectedWidgets[0];
-
   const isCorner = CORNERS.includes(anchor);
   const isSide = SIDES.includes(anchor);
   const cornerClass = isCorner
@@ -34,19 +29,12 @@ const SelectionBoxAnchor: React.FC<SelectionBoxAnchorProps> = ({ anchor }) => {
     : '';
   const anchorClass = `${cornerClass} ${sideClass}`;
 
-  const isTop = anchor === 'top';
-
   return (
-    <>
-      {isTop && selectedWidgets?.length === 1 && (
-        <WidgetActions widget={widget} />
-      )}
-      <div
-        {...gestureable('resize')}
-        {...anchorable(anchor)}
-        className={anchorClass}
-      />
-    </>
+    <div
+      {...gestureable('resize')}
+      {...anchorable(anchor)}
+      className={anchorClass}
+    />
   );
 };
