+++ packages/dashboard/src/components/widgets/tile/tile.css
@@ -4,7 +4,6 @@
   display: flex;
   flex-direction: column;
   height: 100%;
-  overflow: hidden;
 }
 
 .widget-tile-body {
@@ -18,6 +17,10 @@
 }
 
 .widget-tile-header {
+  overflow: hidden;
+}
+
+.widget-tile-header h1 {
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
