+++ packages/dashboard/src/components/widgets/widget.css
@@ -3,6 +3,7 @@
 .widget {
   position: absolute;
   height: 100%;
+  overflow: hidden;
   user-select: none;
   box-sizing: border-box;
   border: var(--selection-border-width) solid transparent;
