+++ packages/dashboard/src/customization/widgets/barChart/component.tsx
@@ -41,10 +41,13 @@ const BarChartWidgetComponent: React.FC<BarChartWidget> = (widget) => {
   const aggregation = getAggregation(widget);
   const significantDigits =
     widgetSignificantDigits ?? dashboardSignificantDigits;
-
-  const size = { width: chartSize.width, height: chartSize.height };
+  // there may be better ways to fix this, i.e. not have -44 and let the chart container  take its parent height,
+  // the problem is that the Resizable component needs a "height" to be provided,
+  // so not entirely sure if we can have a mechanism where the container auto adjusts the height
+  // the 44 is from the widget tile header's height
+  const size = { width: chartSize.width, height: chartSize.height - 44 };
   return (
-    <WidgetTile widget={widget} title={title}>
+    <WidgetTile widget={widget} removeable title={title}>
       <BarChart
         chartSize={size}
         key={key}
