+++ packages/dashboard/src/customization/widgets/components/no-chart-data.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import { render } from '@testing-library/react';
+import { act, render, screen } from '@testing-library/react';
 import NoChartData from './no-chart-data';
 import { default as lineSvgDark } from '../lineScatterChart/line-dark.svg';
 import WidgetTile from '~/components/widgets/tile/tile';
@@ -18,9 +18,9 @@ describe('NoChartData', () => {
         widgets: [MOCK_LINE_CHART_WIDGET],
       },
     });
-    const { getByText, getByAltText } = render(
+    const { getByText, getByAltText, getByTitle } = render(
       <Provider store={store}>
-        <WidgetTile widget={MOCK_LINE_CHART_WIDGET}>
+        <WidgetTile widget={MOCK_LINE_CHART_WIDGET} removeable>
           <NoChartData
             icon={lineSvgDark}
             emptyStateText='Browse and select to add your asset properties in your line widget.'
@@ -34,9 +34,22 @@ describe('NoChartData', () => {
     const emptyStateText = getByText(
       'Browse and select to add your asset properties in your line widget.'
     );
+    const deleteButton = getByTitle('delete widget');
 
     expect(icon).toBeInTheDocument();
     expect(emptyStateText).toBeInTheDocument();
+    expect(deleteButton).toBeInTheDocument();
+
+    act(() => {
+      deleteButton?.click();
+    });
+
+    const deleteBtn = screen.getByText('Delete');
+    expect(deleteBtn).toBeInTheDocument();
+    act(() => {
+      deleteBtn?.click();
+    });
+    expect(store.getState().dashboardConfiguration.widgets).toEqual([]);
   });
 
   it('renders the correct text widget with delete option', () => {
@@ -45,9 +58,9 @@ describe('NoChartData', () => {
         widgets: [MOCK_TEXT_WIDGET],
       },
     });
-    const { getByText } = render(
+    const { getByText, getByTitle } = render(
       <Provider store={store}>
-        <WidgetTile widget={MOCK_TEXT_WIDGET}>
+        <WidgetTile widget={MOCK_TEXT_WIDGET} removeable>
           <StyledTextArea {...MOCK_TEXT_WIDGET} handleSetEdit={() => {}} />
         </WidgetTile>
         ;
@@ -55,7 +68,20 @@ describe('NoChartData', () => {
     );
 
     const textValue = getByText('text content');
+    const deleteButton = getByTitle('delete widget');
 
     expect(textValue).toBeInTheDocument();
+    expect(deleteButton).toBeInTheDocument();
+
+    act(() => {
+      deleteButton?.click();
+    });
+
+    const deleteBtn = screen.getByText('Delete');
+    expect(deleteBtn).toBeInTheDocument();
+    act(() => {
+      deleteBtn?.click();
+    });
+    expect(store.getState().dashboardConfiguration.widgets).toEqual([]);
   });
 });
